functions.php
Changed the code so that the active class only adds to the current page and not to its category.

$active_class = ($item->current || $item->current_item_ancestor) ? 'active' : '';
$nav_link_class = ( $depth > 0 ) ? 'dropdown-item ' : 'nav-link ';
$attributes .= ( $args->walker->has_children ) ? ' class="'. $nav_link_class . $active_class . ' dropdown-toggle" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"' : ' class="'. $nav_link_class . $active_class . '"';
This code only adds the "active" class if the current item is the current page or a direct descendant of it. It no longer checks for the "current_page_parent" or "current-post-ancestor" classes, which are used to indicate the current page's category or a post's ancestor, respectively.